moved Idea object to new file
