+++ lib/2-do-box.js
@@ -1,4 +1,5 @@
-var $ = require("jquery");
+var $ = require('jquery');
+var Idea = require('./idea');
 
 $(document).ready(function() {
   showFirstTen();
@@ -24,13 +25,6 @@ var downQualities = {
   'none' : 'none'
 };
 
-function Idea(title, body){
-  this.title = title;
-  this.body = body;
-  this.id = Date.now();
-  this.quality = 'none';
-  this.completion = 'new';
-}
 
 function renderIdeaListToPage() {
   ideaList.forEach(function(idea) {
